:recycle: refactor : use RunningConst or SocketConst instead of GlobalConst
